feat: Add Sidebar and Toolbar components for enhanced editor layout and functionality
- Implemented Sidebar component for file management, theme toggling, and statistics display.
- Created Toolbar component for text formatting, link management, and additional editor features.
- Introduced useAutoSave hook to handle automatic saving of file states before page unload.
- Developed useEditorLayout hook to manage sidebar and statistics panel resizing.
- Added useEditorSync hook for synchronizing editor content with active file.
- Created useLinkModal hook for managing link insertion and editing within the editor.
