+++ components/tools/extensions/index.ts
@@ -1,0 +1,3 @@
+// Tiptap extensions barrel export
+export { FontSizeExtension } from "./FontSizeExtension";
+export { VisibilityExtension } from "./VisibilityExtension";
